+++ Changelog.md
@@ -11,6 +11,7 @@ Types of changes: New, Improved, Fixed, Removed.
 
 - New: Added zed extension
 - New: Abc command has new flag `--printed` to direct output to the terminal instead of a file
+- New: Added `dense` field label for tune headers
 - Improved: CLI output flag (`-o`) is now optional for abc and pdf commands
 - Improved: Uses strict concurrency for swift 6 language mode